+++ src/Main.java
@@ -1,9 +1,11 @@
 import model.Employee;
+import view.ManagerView;
 
 import java.io.IOException;
 
 public class Main {
     public static void main(String[] args) throws IOException {
         System.out.println("Shipping Management System");
+
     }
 }
